requirements.txt
boto3==1.17.97
botocore==1.20.97
jmespath==0.10.0
python-dateutil==2.8.1
PyYAML==5.4.1
s3transfer==0.4.2
six==1.16.0
urllib3==1.26.5
